observability: collect every v2 metric definition into one file
Metric definitions were spread across three files in the operator
module:

  * operator/internal/observability/recorder.go    (3 operator_controller_*)
  * operator/internal/observability/stretch_recorder.go (5 operator_stretchcluster_*)
  * operator/internal/controller/redpanda/metric_controller.go (4 v2 redpanda_*)

To answer "what metrics does this operator expose?" you had to
grep three files. Consolidating: every prometheus.New* call now
lives in operator/internal/observability/metrics.go, grouped
into three sections (reconcile-health, StretchCluster, Redpanda
v2 CR resource-state) with a single init() that registers all
12 metrics. The redpanda metric_controller.go reconciler imports
the exported vars from the observability package instead of
defining its own.

The v2 metric vars are now exported (Redpandas,
RedpandaDesiredNodes, RedpandaReadyNodes,
RedpandaMisconfiguredClusters) instead of package-private.

recorder.go is reduced to the package docstring (kept as the
canonical place for the package-level godoc).
stretch_recorder.go is reduced to the four RecordStretchCluster*
helper functions.

Out of scope:

  * v1 (vectorized.redpanda.com Cluster) metrics — that
    controller is legacy and explicitly excluded from
    unrelated changes. Its 4 metrics remain defined next to
    its reconciler.
  * Multicluster raft metrics in pkg/multicluster/leaderelection/
    metrics.go — different Go module; cross-module consolidation
    would either expose operator/internal/observability/ to all
    pkg/ consumers or pull controller-runtime deps into the pkg
    module.

No emitted metric names change. No alerts or dashboards affected.
